requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.5
asyncpg==0.29.0
sqlalchemy==2.0.29
python-dotenv==1.0.1